Format and tidy code
